+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/chat/GroupChatImpl.kt
@@ -1,19 +1,15 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.chat
 
-import com.github.insanusmokrassar.TelegramBotAPI.types.*
-import com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.*
+import com.github.insanusmokrassar.TelegramBotAPI.types.ChatId
+import com.github.insanusmokrassar.TelegramBotAPI.types.ChatPhoto
 import com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.GroupChat
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.RawMessage
 
-data class SupergroupChat(
+data class GroupChatImpl(
     override val id: ChatId,
     override val title: String? = null,
-    override val username: Username? = null,
-    override val description: String? = null,
     override val allMembersAreAdmins: Boolean,
     override val inviteLink: String? = null,
     override val chatPhoto: ChatPhoto? = null,
-    override val pinnedMessage: RawMessage? = null,
-    val stickerSetName: String? = null,
-    val canSetStickerSet: Boolean
-) : GroupChat, UsernameChat, DescriptionChat
+    override val pinnedMessage: RawMessage? = null
+) : GroupChat