+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/chat/PrivateChat.kt
@@ -1,11 +1,13 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.chat
 
 import com.github.insanusmokrassar.TelegramBotAPI.types.*
+import com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.Chat
+import com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.UsernameChat
 
 data class PrivateChat(
     override val id: ChatId,
-    val username: Username? = null,
+    override val username: Username? = null,
     val firstName: String? = null,
     val lastName: String? = null,
     override val chatPhoto: ChatPhoto? = null
-) : Chat
+) : Chat, UsernameChat