+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/chat/abstracts/Chat.kt
@@ -1,0 +1,9 @@
+package com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts
+
+import com.github.insanusmokrassar.TelegramBotAPI.types.ChatId
+import com.github.insanusmokrassar.TelegramBotAPI.types.ChatPhoto
+
+interface Chat {
+    val id: ChatId
+    val chatPhoto: ChatPhoto?
+}