+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/chat/abstracts/DescriptionChat.kt
@@ -1,0 +1,5 @@
+package com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts
+
+interface DescriptionChat : PublicChat {
+    val description: String?
+}