+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/chat/abstracts/UsernameChat.kt
@@ -1,0 +1,7 @@
+package com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts
+
+import com.github.insanusmokrassar.TelegramBotAPI.types.Username
+
+interface UsernameChat : Chat {
+    val username: Username?
+}